+++ src/browsers.ts
@@ -1,13 +1,13 @@
 import puppeteer, { Browser } from "puppeteer"
 
-let chrome: Browser | null = null
-let firefox: Browser | null = null
+let chrome: Promise<Browser> | null = null
+let firefox: Promise<Browser> | null = null
 
 async function getChromium ( ): Promise<Browser | null> {
   if ( chrome ) return chrome
 
   try {
-    chrome = await puppeteer.launch ( {
+    chrome = puppeteer.launch ( {
       args: [ "--no-sandbox", "--disable-setuid-sandbox" ],  // Useful in certain environments (e.g., Docker)
     } )
   } catch ( err ) {
@@ -21,7 +21,7 @@ async function getFirefox ( ): Promise<Browser | null> {
   if ( firefox ) return firefox
 
   try {
-    firefox = await puppeteer.launch ( {
+    firefox = puppeteer.launch ( {
       browser: "firefox", // Use Firefox
       headless: true, // Set to false if you want to see the browser
       args: [ "--no-sandbox", "--disable-setuid-sandbox" ],  // Useful in certain environments (e.g., Docker)
@@ -33,22 +33,26 @@ async function getFirefox ( ): Promise<Browser | null> {
   return firefox
 }
 
-// Function to close both browsers
 async function closeBrowsers ( ): Promise<void> {
   try {
     if ( chrome ) {
-      await chrome.close ( )
+      await ( await chrome ).close ( )
       chrome = null
       console.log ( "Chromium Browser Closed" )
     }
     if ( firefox ) {
-      await firefox.close ( )
+      await ( await firefox ).close ( )
       firefox = null
       console.log ( "Firefox Browser Closed" )
     }
   } catch ( err ) {
     console.error ( "Error closing browsers in @iqx-limited/quick-form:", err )
   }
+}
+
+async function launchBrowsers ( ): Promise<void> {
+  await getChromium ( )
+  await getFirefox ( )
 }
 
 // Listen for process exit and close browsers
@@ -68,4 +72,4 @@ process.on ( "SIGTERM", async ( ) => {
 } )
 
 // Export the browsers for use in other files
-export { getChromium, getFirefox, closeBrowsers }
+export { getChromium, getFirefox, closeBrowsers, launchBrowsers }