+++ src/index.ts
@@ -1,4 +1,5 @@
 import { satisfies } from "semver"
+import { launchBrowsers } from "./browsers.js"
 
 const requiredVersion = ">=20.0.0"
 
@@ -8,6 +9,9 @@ if ( !satisfies ( process.version, requiredVersion ) ) {
   )
   process.exit ( 1 )
 }
+
+// Start the browsers
+launchBrowsers ( )
 
 export * from "./modules/pdf2img.js"