+++ CHANGELOG.md
@@ -2,8 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+# 1.3.2 (2025-02-09)
+- Launch Browsers Early
+- Allow manually launching the browsers
+
 # 1.3.1 (2025-02-09)
-= Bug Fixes
+- Bug Fixes
 - Export Close Browsers
 
 # 1.3.0 (2025-02-09)